pyproject.toml
[project]
name = "fastpantry"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "bs4>=0.0.2",
    "python-dateutil>=2.9.0.post0",
    "requests>=2.32.3",
]
